Move build from workflow job to container action
I've changed the setup of the natives-linux job to not require needing Node 20, so hopefully it will continue to work with newer versions of checkout and upload-artifact for the forseeable future. I've done this by moving the docker setup into its own action (which is local to the repo) that runs a script in the container that builds the natives and snapshots, but doesn't use any third party actions - meaning it should not be need node 20 to be used in the container.

Other minor changes:

- Used apt-get instead of apt because using apt in a script gives the warning message "apt does not have a stable interface".
- Removed usages of sudo just to be consistent (I can't always use sudo because initially sudo isn't installed)
- Added --no-daemon to gradle command because only one gradle command is being run so the daemon isn't necessary.
- Added --quiet to apt-get usages to cut down unuseful lines in logs
- Moved around arguments to apt-get so they're always at the front
- removed setup-java and gradle-build-action usages in the natives-linux step because zulu installation is done in the container and gradle is downloaded when the wrapper is run.
- gnupg gets installed anyway as a dependency but I added it explicitly because it is needed for dearmoring the azul key

Having the action be in a separate actions folder may not be desirable, in which case I can probably move the action into the workflow folder somewhere but I'm not sure where.

Copied docker action structure from my similar PR in Jamepad (https://github.com/libgdx/Jamepad/pull/34)