Bugfix, a few more minor opts in DataOutputStream.
Double.doubleToLongBits() can be replaced with the better alternative from Numbers, which is already used for writing floats. Shifts by long values act the same as shifts by int values, but allocating a long is slow on GWT when we only need an int, so that's changed in writeLong() . In writeUTF(), all other comparisons are done on ints, so it just seems more fitting to compare with 0x800 than '\u0800' (which might actually be a String in JS? I am not sure...). There was also a more-serious bug in writeUTF() because it compared with 80 instead of 0x80 (or decimal 128), which would have made this output invalid UTF8 sequences. In writeBytes(), calling os.write() already does a mask to use the lowest 8 bits only, so we don't need to redo it here.