Two GWT emu fixes re: long in DataInputStream. (#7590)
* Two GWT emu fixes re: long in DataInputStream.

The first change is just using the added-a-few-years-ago Numbers.longBitsToDouble(), which is faster, instead of GWT's slow Double.longBitsToDouble() . There's an extended discussion (with pictures!) of the second change here, https://tommyettinger.github.io/libGDX-GWT-checks/ . The reason the second change looks the way it does now is optimistically hoping that JS will be able to pipeline two non-dependent bit operations before it ORs them. This mostly affects UBJson, but it may affect other places that used DataInputSteam.readLong() on GWT. Long story short, an `L` was missing from a mask literal, and so it wasn't masking an int at all, letting it implicitly promote to long afterwards and sign-extending then if the int was negative.

* Bugfix, a few more minor opts in DataOutputStream.

Double.doubleToLongBits() can be replaced with the better alternative from Numbers, which is already used for writing floats. Shifts by long values act the same as shifts by int values, but allocating a long is slow on GWT when we only need an int, so that's changed in writeLong() . In writeUTF(), all other comparisons are done on ints, so it just seems more fitting to compare with 0x800 than '\u0800' (which might actually be a String in JS? I am not sure...). There was also a more-serious bug in writeUTF() because it compared with 80 instead of 0x80 (or decimal 128), which would have made this output invalid UTF8 sequences. In writeBytes(), calling os.write() already does a mask to use the lowest 8 bits only, so we don't need to redo it here.