Updated BaseTmjMapLoader to change a private method to protected to Match the BaseTmxMapLoader. Needed the ability to override this method anyway.
Added support for animations when using Collection of Image Tilesets.

Added new test map which shows off collection of image tileset as with animations.

Added a flag to ignore processing of tilesets which use a collection of images, if the user may have some reason to do.

Added a new overloaded processInputDir() method meant for external use.

Updated Test's with the new map.