TiledMapPacker Tool Enhancement, ImageLayer and Json map Support, + crashfix. (#7525)
* Added functionality to the TiledMapPacker to support image layers.
Modified AtlasTmxMapLoader to not crash when handling imagelayers.

* Added .tmj parsing for TiledMapPacker.
Added AtlasTmjMapLoader
Changed check in getImage() AtlasResolver

* Cleaned up and added comments referring to .tmj files.
Added a task to build a runnable for the tiledmappacker in the build.gradle

* Added new map file with image layers for atlas packing test.
Added processed map to tiled-atlas-processed folder.
Modified TiledMapAtlasAssetManagerTest to toggle between both test maps.

* Apply formatter

* Added support for loading maps which use custom classes but allowing a new command line argument to pass in the path to the tiled project file
Changed how command line parameters are passed in. I cleaned it up a bit to allow for more flexibility with the command line arguments. It was kind of broken originally. The options only worked properly if an output folder parameter was added.

* Apply formatter

* Changes to TiledMapPackerTest:
Fixed incorrect paths.
Added additional parameter options to support updates to the TiledMapPacker.
There are now up to 4 different maps that can be generated after running TiledMapPackerTest. To show off more variety of map times the TiledMapPacker can process.

Changes to TiledMapPackerTestRender:
Fixed incorrect path.
A user can choose to test up to 4 different tiled maps that are generated in the deleteMe folder by changing an enum.

Updated maps that are available in the tiled-atlas-processed folder.
These maps are always viewable through the newly added TiledMapJsonAtlasAssetManagerTest and the original TiledMapAtlasAssetManagerTest.

Changes to TiledMapPacker:
Removed using UUID to generate unique image names. Now we simply use a System.nanoTime() unique id counter.

* Apply formatter

* Added Support for Packing a map when using a tileset made up of a "Collection of Images"

Fixed image layers not showing up because they were in a group layer.

* Apply formatter

* Updated BaseTmjMapLoader to change a private method to protected to Match the BaseTmxMapLoader. Needed the ability to override this method anyway.

Added support for animations when using Collection of Image Tilesets.

Added new test map which shows off collection of image tileset as with animations.

Added a flag to ignore processing of tilesets which use a collection of images, if the user may have some reason to do.

Added a new overloaded processInputDir() method meant for external use.

Updated Test's with the new map.

* Apply formatter

* Cleaned up a few comments/typos

* Apply formatter

---------

Co-authored-by: GitHub Action <action@github.com>