Dubious? Avoid heap pollution in appendAll().
This requires making `appendAll(T...)` final, because the JVM can apparently get "heap pollution" when generic-typed varargs are used by a method that isn't at least one of private, static, and/or final. I haven't found any documentation on what "heap pollution" entails, so this might not be needed. I chose not to apply the SafeVarargs annotation because it isn't used anywhere else in libGDX, even though it is a valid compiler warning to omit it. The check for non-zero length before wasn't really... helpful, I think, since the loop just wouldn't run if array was empty. But, the parameter could be null in some situations, and iterating over that would crash with an NPE, so now it checks that the parameter is not null.