Changed the root.get("tileSets") to root.get("tilesets") in TmjMapLoader.java
Originally JsonValue.get() ignored case. So it worked accidentally. But recently it was changed. Checking against "tileSets" was a mistake to begin with so it needs to be fixed. The TMX map loader IS case-sensitive, so it should match. Which is why I am not suggesting to use the getIgnoreCase() method that was added.