For issue #7024
You can now create HexagonalTiledMapRenderer objects when passing in new empty TiledMap objects. This will no longer result in a crash and allows similar functionality to all the other TiledMapRenderers.
Also added getters and setters for the necessary fields which would be needed for a user to set it up manually.