Clean-up in CharArray. (#7711)
* Clean up docs in CharArray.

Dangling comment here, typo there. The only substantive changes here are avoiding a duplicate check for equal size in equals() and avoiding an explicit call to Long.toString() that I don't think is required (it's already being appended to a String).

* Dubious? Avoid heap pollution in appendAll().

This requires making `appendAll(T...)` final, because the JVM can apparently get "heap pollution" when generic-typed varargs are used by a method that isn't at least one of private, static, and/or final. I haven't found any documentation on what "heap pollution" entails, so this might not be needed. I chose not to apply the SafeVarargs annotation because it isn't used anywhere else in libGDX, even though it is a valid compiler warning to omit it. The check for non-zero length before wasn't really... helpful, I think, since the loop just wouldn't run if array was empty. But, the parameter could be null in some situations, and iterating over that would crash with an NPE, so now it checks that the parameter is not null.

* The method is writer(), not asWriter().

* CharArray shouldn't behave differently from...

... every other Array type here; none of the others throw Exceptions when the Array is empty.