For issue #7726
There was an issue with Templates not taking into account the proper GIDs when multiple tilesets were used. Implemented the fixes suggested by user who submitted the issue.

Modified both the Tmj and Tmx map loaders to keep them in sync. And updated the template test to use a template which references another tileset, which would break if this fix was not put in place. The template test now uses textures from 2 different tilesets.